Added reconnection failover. Standardized naming.
Added reconnection failover for collapsed tunnels. Tunnels that cannot be
restarted after 5 retries are now changed to the first entry in
"backup_servers".  Behavior is of a FIFO, "default_server" is pushed to
the end of "backup_servers" and the first entry of "backup_servers" is
popped into "default_server"

Standardized naming scheme from camelCase to using_underscores.

Made README.md acutally use markdown.  Added the goal of using the script
without super-user.  still need to investigate best option for
implementing this.